Move special functionality from hue_api.ts to dashboard.ts
