+++ src/hue_api.ts
@@ -1,6 +1,5 @@
 import http from 'http';
 import {Groups, Light, Lights, Scene, Scenes} from './hue_api_types';
-import * as Conversions from './conversions';
 
 export const TRANSITION_TIME_UNITS_PER_SECOND = 10;
 export const TRANSITION_TIME_SECONDS_DEFAULT = 0.4;
@@ -48,29 +47,3 @@ export const request = (method, path, body) => new Promise((resolve, reject) => 
   req.write(JSON.stringify(body));
   req.end();
 });
-
-export const getRoomsFromGroups = (groups: Groups) => {
-  const rooms = [];
-
-  for(const groupId in groups) {
-    const group = groups[groupId];
-    if(group.type == 'Room' || group.type == 'Zone') {
-      let colour = "";
-      if(group.state.any_on) {
-        if(group.action.xy && group.action.bri) {
-          colour = Conversions.xyBriToHex(group.action.xy[0], group.action.xy[1], group.action.bri);
-        } else if(group.action.colormode == 'ct') {
-          colour = Conversions.ctToHex(group.action.ct);
-        }
-      } else colour = "#000000";
-      rooms.push({
-        id: groupId,
-        name: group.name,
-        state: group.state,
-        colour: colour
-      });
-    }
-  }
-
-  return rooms;
-};
